Rearrange codes and address comments
