+++ Codes/data_prep/config.yaml
@@ -3,3 +3,5 @@
 lab_lookback_window: 5 # days
 symp_lookback_window: 30 # days
 ed_visit_lookback_window: 5 # years
+align_on: 'treatment-dates' 
+main_date_col: 'treatment_date'
