--- Codes/data_prep/constants_postprocess.py
@@ -1,65 +1,0 @@
-symp_cols = [
-    'esas_pain',
-    'esas_tiredness',
-    'esas_nausea',
-    'esas_depression',
-    'esas_anxiety',
-    'esas_drowsiness',
-    'esas_appetite',
-    'esas_well_being',
-    'esas_shortness_of_breath',
-    'patient_ecog',
-]
-lab_cols = [
-    'alanine_aminotransferase',
-    'albumin',
-    'alkaline_phosphatase',
-    'aspartate_aminotransferase',
-    'basophil',
-    'bicarbonate',
-    'chloride',
-    'creatinine',
-    'eosinophil',
-    'glucose',
-    'hematocrit',
-    'hemoglobin',
-    'lactate_dehydrogenase',
-    'lymphocyte',
-    'magnesium',
-    'mean_corpuscular_hemoglobin',
-    'mean_corpuscular_hemoglobin_concentration',
-    'mean_corpuscular_volume',
-    'mean_platelet_volume',
-    'monocyte',
-    'neutrophil',
-    'phosphate',
-    'platelet',
-    'potassium',
-    'red_blood_cell',
-    'red_cell_distribution_width',
-    'sodium',
-    'total_bilirubin',
-    'white_blood_cell',
-]
-symp_change_cols = [f'{col}_change' for col in symp_cols]
-lab_change_cols = [f'{col}_change' for col in lab_cols]
-drug_cols = [
-    '%_ideal_dose_given_CISPLATIN',
-    '%_ideal_dose_given_FLUOROURACIL', 
-    '%_ideal_dose_given_ETOPOSIDE',
-    '%_ideal_dose_given_CARBOPLATIN', 
-    '%_ideal_dose_given_OXALIPLATIN',
-    '%_ideal_dose_given_PEMBROLIZUMAB', 
-    '%_ideal_dose_given_NIVOLUMAB',
-    '%_ideal_dose_given_DOCETAXEL', 
-    '%_ideal_dose_given_CYCLOPHOSPHAMIDE',
-    '%_ideal_dose_given_DURVALUMAB', 
-    '%_ideal_dose_given_CETUXIMAB',
-    '%_ideal_dose_given_RAMUCIRUMAB', 
-    '%_ideal_dose_given_RALTITREXED',
-    '%_ideal_dose_given_IPILIMUMAB', 
-    '%_ideal_dose_given_MITOMYCIN',
-    '%_ideal_dose_given_PANITUMUMAB', 
-    '%_ideal_dose_given_CAPECITABINE',
-    '%_ideal_dose_given_ERLOTINIB',
-]
