+++ Codes/data_prep/preprocess/dart.py
@@ -1,9 +1,9 @@
 """
 Module to preprocess DART (symptom data)
 """
-from typing import Optional
+# from typing import Optional
 import pandas as pd
-import numpy as np
+# import numpy as np
 
 
 def get_symptoms_data(esas_data_file):
@@ -17,12 +17,18 @@ def get_symptoms_data(esas_data_file):
 
 def process_symptoms_data(df):
     
-    # get indices of ecog scores containing values with descriptions
-    ecog_locs = np.where(df['patient_ecog'].apply(lambda x: len(str(x)) > 3))[0]
-    if len(ecog_locs)>0:
-        for i1 in range(len(ecog_locs)):
-            df['patient_ecog'].iloc[ecog_locs[i1]] = df['patient_ecog'].iloc[ecog_locs[i1]][0]
+    # # get indices of ecog scores containing values with descriptions
+    # ecog_locs = np.where(df['patient_ecog'].apply(lambda x: len(str(x)) > 3))[0]
+    # if len(ecog_locs)>0:
+    #     for i1 in range(len(ecog_locs)):
+    #         df['patient_ecog'].iloc[ecog_locs[i1]] = df['patient_ecog'].iloc[ecog_locs[i1]][0]
      
+    # df['patient_ecog'] = pd.to_numeric(df['patient_ecog'])
+    
+    # some patient_ecog entries have the following format: score-description
+    # remove the descriptions and convert the scores to int
+    mask = df['patient_ecog'].str.contains('-')
+    df.loc[mask, 'patient_ecog'] = df.loc[mask, 'patient_ecog'].str.split('-').str[0]
     df['patient_ecog'] = pd.to_numeric(df['patient_ecog'])
     
     # order by survey date
